+++ src/main/kotlin/br/com/gazintech/plugin/spring_utils/api/idempotency/repository/IdempotencyCache.kt
@@ -1,11 +1,10 @@
-package br.com.gazintech.plugin.spring_utils.api.idempotency
+package br.com.gazintech.plugin.spring_utils.api.idempotency.repository
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import org.springframework.data.redis.core.RedisHash
 import org.springframework.data.redis.core.TimeToLive
 import java.io.Serializable
-import java.util.*
-
+import java.util.UUID
 
 @RedisHash(value = "idempotency-cache")
 class IdempotencyCache(
