+++ src/main/kotlin/br/com/gazintech/plugin/spring_utils/api/idempotency/repository/IdempotencyRepository.kt
@@ -1,5 +1,6 @@
-package br.com.gazintech.plugin.spring_utils.api.idempotency
+package br.com.gazintech.plugin.spring_utils.api.idempotency.repository
 
+import br.com.gazintech.plugin.spring_utils.api.idempotency.repository.IdempotencyCache
 import java.util.*
